230111 14:42 공지 수정, 문의 글쓰기/수정/삭제

+++ boardGameCafe/src/main/java/kr/notice/action/NoticeDeleteAction.java
@@ -2,6 +2,7 @@ package kr.notice.action;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import kr.controller.Action;
 import kr.notice.dao.NoticeDAO;
@@ -11,11 +12,24 @@ import kr.util.FileUtil;
 public class NoticeDeleteAction implements Action {
 	@Override
 	public String execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
+		HttpSession session = request.getSession();
+		
+		Integer user_num = (Integer)session.getAttribute("user_num");
+		Integer user_auth = (Integer)session.getAttribute("user_auth");
+		
+		if(user_num == null) {//로그인이 되지 않은 경우
+			return "redirect:/member/loginForm.do";
+		}
+		
 		int noti_num = Integer.parseInt(request.getParameter("noti_num"));
 		NoticeDAO dao = NoticeDAO.getInstance();
 		NoticeVO notice = dao.getNotice(noti_num);
 		
-		//로그인한 회원번호와 작성자 회원번호가 일치
+		if(user_auth != 9) {
+			//로그인한 회원번호와 작성자 회원번호가 불일치
+			return "/WEB-INF/views/notice/noticeList.jsp";
+		}
+		
 		dao.deleteNotice(noti_num);
 		
 		//파일 삭제
